refactor: reorganize CRM components and improve industry options handling
