+++ frontend/src/community/crm/v2/components/templates/SidePanelWrapper/SidePanelWrapper.tsx
