+++ frontend/src/community/crm/v2/constants/commonConstants.ts
@@ -1,5 +1,1 @@
-export const PHONE_NUMBER_PREFIX = "+";
-export const CURRENCY_PREFIX = "$";
-export const EMPTY_PLACEHOLDER = "-";
-
 export const COMPANY_PAGE_SIZE = 10;
