+++ frontend/src/community/crm/v2/utils/commonUtil.ts
@@ -1,21 +1,14 @@
-import {
-  CURRENCY_PREFIX,
-  EMPTY_PLACEHOLDER
-} from "~community/crm/v2/constants/commonConstants";
-
 const isEmptyValue = (value?: string | number) =>
   value === undefined || Number(value) === 0;
 
 export const formatTableValue = (value?: string | number, prefix = "") =>
-  isEmptyValue(value) ? EMPTY_PLACEHOLDER : `${prefix}${value}`;
+  isEmptyValue(value) ? "-" : `${prefix}${value}`;
 
 export const formatMonetaryValue = (value?: string) => {
-  if (isEmptyValue(value)) return EMPTY_PLACEHOLDER;
+  if (isEmptyValue(value)) return "-";
 
-  return `${CURRENCY_PREFIX}${value?.split(".")[0]}`;
+  return `$${value?.split(".")[0]}`;
 };
 
 export const formatMonetaryValueWithDecimals = (value?: string | number) =>
-  isEmptyValue(value)
-    ? EMPTY_PLACEHOLDER
-    : `${CURRENCY_PREFIX}${Number(value).toFixed(2)}`;
+  isEmptyValue(value) ? "-" : `$${Number(value).toFixed(2)}`;
